// src/aws/interfaces/IFilm.ts
export default interface IFilm {
  actors?: string;
  company?: string;
  countries?: string;
  description: string;
  directors?: string;
  duration?: number;
  genres?: string;
  imdbID: number;
  languages?: string;
  rating?: number;
  ratingPercentile?: number;
  releaseYear: number;
  title: string;
  writers?: string;
}